- this time the signed/unsigned mismatches are really solveD

+++ packages/data/c/src/cvec.c
@@ -371,7 +371,7 @@ cvec_fft_fvec(fts_object_t *o, int winlet, fts_symbol_t s, int ac, const fts_ato
   int in_size = fvec_get_size(in);
   unsigned int fft_size = fts_get_fft_size(in_size);
   float *fft_ptr;
-  unsigned int i = 0;
+  int i = 0;
 
   cvec_set_size(this, fft_size >> 1);
   fft_ptr = (float *)cvec_get_ptr(this);
@@ -380,7 +380,7 @@ cvec_fft_fvec(fts_object_t *o, int winlet, fts_symbol_t s, int ac, const fts_ato
     fft_ptr[i] = in_ptr[i];
 
   /* zero padding */
-  for(; i<fft_size; i++)
+  for(; i< (int)fft_size; i++)
     fft_ptr[i] = 0.0;
 
   fts_rfft_inplc(fft_ptr, fft_size);
@@ -404,13 +404,13 @@ cvec_fft_cvec(fts_object_t *o, int winlet, fts_symbol_t s, int ac, const fts_ato
   if(in != this)
   {
     complex *in_ptr = cvec_get_ptr(in);
-    unsigned int i;
+    int i;
 
     for(i=0; i<in_size; i++)
       fft_ptr[i] = in_ptr[i];
 
     /* zero padding */
-    for(; i<fft_size; i++)
+    for(; i< (int)fft_size; i++)
       fft_ptr[i] = CZERO;
   }
 
@@ -447,13 +447,13 @@ cvec_ifft_cvec(fts_object_t *o, int winlet, fts_symbol_t s, int ac, const fts_at
   if(in != this)
   {
     complex *in_ptr = cvec_get_ptr(in);
-    unsigned int i;
+    int i;
 
     for(i=0; i<in_size; i++)
       fft_ptr[i] = in_ptr[i];
 
     /* zero padding */
-    for(; i<fft_size; i++)
+    for(; i< (int)fft_size; i++)
       fft_ptr[i] = CZERO;
   }
